move commands into seperate packages (#35)
